refactor : BlackPawnStrategy와 WhitePawnStrategy를 PawnStrategy로 추상화
